+++ front/src/components/certificates/CertificateCard.js
@@ -1,21 +1,28 @@
 import React from "react";
-import * as Api from "../../api";
 import { Card, Row, Button, Col } from "react-bootstrap";
+import * as Api from "../../api";
 import * as dateFns from "date-fns";
+import Swal from "sweetalert2";
 
-function CertificateCard({
-  certificate,
-  setIsEditing,
-  isEditable,
-  getCertificates,
-}) {
-  // const t = new Date(certificate.acquiredAt);
-  // const time_text = t.toISOString().split("T")[0];
+const CertificateCard = ({ certificate, setIsEditing, isEditable, getCertificates }) => {
   const handleDelete = async () => {
     try {
-      await Api.delete("certificates", certificate.certificateId);
-      const res = await Api.get("certificates", certificate["userId"]);
-      getCertificates(res.data);
+      Swal.fire ({
+        icon: "warning",
+        title: "정말 삭제하시겠습니까?",
+        text: "삭제하면 다시 복구할 수 없습니다.",
+        showCancelButton: true
+      }).then(async (result) => {
+        if(result.isConfirmed) {
+          await Api.delete("certificates", certificate.certificateId);
+          const res = await Api.get("certificates", certificate["userId"]);
+          getCertificates(res.data);
+          Swal.fire ({
+            title: "삭제가 완료되었습니다!",
+            icon: "success"
+          })
+        }
+      })
     } catch (e) {
       console.log(e);
     }
@@ -35,19 +42,16 @@ function CertificateCard({
         {isEditable && (
           <Col xs lg="2">
             <Button
+              className="me-2"
               variant="outline-info"
               size="sm"
               onClick={() => setIsEditing(true)}
-            >
-              Edit
-            </Button>{" "}
+            >Edit</Button>
             <Button
               variant="outline-secondary"
               size="sm"
               onClick={() => handleDelete()}
-            >
-              Delete
-            </Button>
+            >Delete</Button>
           </Col>
         )}
       </Row>
