+++ front/src/components/project/ProjectCard.js
@@ -2,13 +2,28 @@ import React from "react";
 import { Card, Row, Button, Col } from "react-bootstrap";
 import * as dateFns from "date-fns";
 import * as Api from "../../api";
+import Swal from "sweetalert2";
+
 
 const ProjectCard = ({ setIsEditing, isEditable, project, setProjects }) => {
   const handleDelete = async () => {
     try {
-      await Api.delete("projects", project.projectId);
-      const res = await Api.get("projects", project["userId"]);
-      setProjects(res.data);
+      Swal.fire ({
+        icon: "warning",
+        title: "정말 삭제하시겠습니까?",
+        text: "삭제하면 다시 복구할 수 없습니다.",
+        showCancelButton: true
+      }).then(async (result) => {
+        if(result.isConfirmed) {
+          await Api.delete("projects", project.projectId);
+          const res = await Api.get("projects", project["userId"]);
+          setProjects(res.data);
+          Swal.fire ({
+            title: "삭제가 완료되었습니다!",
+            icon: "success"
+          })
+        }
+      })
     } catch (err) {
       console.log("삭제가 정상적으로 이루어지지 않았습니다.", err);
     }
@@ -30,19 +45,16 @@ const ProjectCard = ({ setIsEditing, isEditable, project, setProjects }) => {
         {isEditable && (
           <Col xs lg="2">
             <Button
+              className="me-2"
               variant="outline-info"
               size="sm"
               onClick={() => setIsEditing((prev) => !prev)}
-            >
-              Edit
-            </Button>{" "}
+            >Edit</Button>
             <Button
               variant="outline-secondary"
               size="sm"
               onClick={() => handleDelete()}
-            >
-              Delete
-            </Button>
+            >Delete</Button>
           </Col>
         )}
       </Row>
